@@ -13,47 +13,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.0.0",
-    "@angular/cdk": "^9.1.2",
-    "@angular/cli": "~9.0.1",
-    "@angular/common": "~9.0.0",
-    "@angular/compiler": "~9.0.0",
-    "@angular/compiler-cli": "~9.0.0",
-    "@angular/core": "~9.0.0",
-    "@angular/flex-layout": "^9.0.0-beta.29",
-    "@angular/forms": "~9.0.0",
-    "@angular/material": "^9.1.2",
-    "@angular/platform-browser": "~9.0.0",
-    "@angular/platform-browser-dynamic": "~9.0.0",
-    "@angular/router": "~9.0.0",
+    "@angular/animations": "~11.1.0-next.0",
+    "@angular/cdk": "^11.0.0",
+    "@angular/cli": "~11.1.0-next.0",
+    "@angular/common": "~11.1.0-next.0",
+    "@angular/compiler": "~11.1.0-next.0",
+    "@angular/compiler-cli": "~11.1.0-next.0",
+    "@angular/core": "~11.1.0-next.0",
+    "@angular/flex-layout": "^11.0.0-beta.33",
+    "@angular/forms": "~11.1.0-next.0",
+    "@angular/material": "^11.0.0",
+    "@angular/platform-browser": "~11.1.0-next.0",
+    "@angular/platform-browser-dynamic": "~11.1.0-next.0",
+    "@angular/router": "~11.1.0-next.0",
     "@primer/octicons": "^9.6.0",
     "express": "^4.17.1",
     "path": "^0.12.7",
     "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
-    "typescript": "~3.7.5",
-    "zone.js": "~0.10.2"
+    "tslib": "^2.0.0",
+    "typescript": "~4.0.5",
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.1",
-    "@angular/cli": "~9.0.1",
-    "@angular/compiler-cli": "~9.0.0",
-    "@angular/language-service": "~9.0.0",
+    "@angular-devkit/build-angular": "~0.1101.0-next.0",
+    "@angular/language-service": "~11.1.0-next.0",
     "@types/node": "^12.11.1",
-    "@types/jasmine": "~3.5.0",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.1.2",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.3.0",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.2.3",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "protractor": "~5.4.3",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "tslint": "~5.18.0",
-    "typescript": "~3.7.5"
+    "tslint": "~6.1.0"
   },
   "engines": {
     "node": "12.15.0",
